removed old hook implementation of auth and moved to helper function from context
