+++ src/context/AuthProvider.tsx
@@ -20,7 +20,13 @@ export function AuthProvider({ children }: AuthProviderProps) {
 }
 
 export function useAuth() {
-    return useContext(AuthContext);
+    const context = useContext(AuthContext);
+
+    if (context === undefined) {
+        throw new Error("useAuth must be used within an AuthProvider");
+    }
+
+    return context;
 }
